Honor fractional archive stream time budgets
to_i turned a 50ms test budget into unlimited. Compare seconds as
floats so Timeout and 7z reads can fire under one second.

+++ api/app/services/archive_member_streamer.rb
@@ -130,7 +130,7 @@ class ArchiveMemberStreamer
   end
 
   def with_time_budget(max_seconds)
-    limit = max_seconds.to_i
+    limit = max_seconds.to_f
     return yield if limit <= 0
 
     Timeout.timeout(limit) { yield }
